+++ tests/test_frontend/conftest.py
@@ -335,3 +335,18 @@ def patched_clean_document(monkeypatch):
         return original_clean_document(self)
 
     monkeypatch.setattr(UploadDocumentsForm, "clean_document", mock_clean_document)
+
+
+@pytest.fixture(autouse=True, scope="function")
+def bypass_login(monkeypatch, test_apply_user):
+    """Overrides the get_user function to always return the public user without authentication.
+
+    Effectively bypasses the login process for all frontend tests.
+    """
+
+    def patched_get_user(request):
+        return test_apply_user
+
+    monkeypatch.setattr("django.contrib.auth.get_user", patched_get_user)
+
+    yield
